Move Workflow into package with all the other parsing classes
