+++ src/main/java/demo/TemplateParseDemo.java
@@ -15,8 +15,8 @@ import org.opensearch.client.node.NodeClient;
 import org.opensearch.common.SuppressForbidden;
 import org.opensearch.common.io.PathUtils;
 import org.opensearch.flowframework.template.Template;
+import org.opensearch.flowframework.template.Workflow;
 import org.opensearch.flowframework.template.WorkflowProcessSorter;
-import org.opensearch.flowframework.workflow.Workflow;
 import org.opensearch.flowframework.workflow.WorkflowStepFactory;
 
 import java.io.IOException;
