+++ src/main/java/org/opensearch/flowframework/template/Workflow.java
@@ -6,14 +6,12 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.flowframework.workflow;
+package org.opensearch.flowframework.template;
 
 import org.opensearch.core.xcontent.ToXContentObject;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.core.xcontent.XContentParser;
-import org.opensearch.flowframework.template.Template;
-import org.opensearch.flowframework.template.WorkflowEdge;
-import org.opensearch.flowframework.template.WorkflowNode;
+import org.opensearch.flowframework.workflow.WorkflowData;
 
 import java.io.IOException;
 import java.util.ArrayList;
