+++ src/main/java/org/opensearch/flowframework/template/WorkflowProcessSorter.java
@@ -10,7 +10,6 @@ package org.opensearch.flowframework.template;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.opensearch.flowframework.workflow.Workflow;
 import org.opensearch.flowframework.workflow.WorkflowData;
 import org.opensearch.flowframework.workflow.WorkflowStep;
 import org.opensearch.flowframework.workflow.WorkflowStepFactory;
