+++ src/test/java/org/opensearch/flowframework/template/TemplateTestJsonUtil.java
@@ -14,7 +14,6 @@ import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.core.xcontent.ToXContentObject;
 import org.opensearch.core.xcontent.XContentParser;
-import org.opensearch.flowframework.workflow.Workflow;
 
 import java.io.IOException;
 import java.util.List;
